Chart report code improvements and refactoring
