--- modern/src/common/chartTypes.js
@@ -1,7 +1,0 @@
-import t from '../common/localization';
-
-export const chartTypes = [
-  { id: 'speed', name: t('positionSpeed') },
-  { id: 'accuracy', name: t('positionAccuracy') },
-  { id: 'altitude', name: t('positionAltitude') },
-];
