+++ modern/src/common/converter.js
@@ -1,0 +1,24 @@
+const speedConverter = (value, unit) => {
+  let factor;
+  switch (unit) {
+    case 'kmh':
+      factor = 1.852;
+    case 'mph':
+      factor = 1.15078;
+    case 'kn':
+    default:
+      factor = 1;
+  }
+  return (value * factor).toFixed(2);
+};
+
+export const getConverter = (key) => {
+  switch (key) {
+    case 'speed':
+      return speedConverter;
+    default:
+      return function (value) {
+        return value;
+      }
+  }
+};
