+++ modern/src/common/formatter.js
@@ -80,26 +80,3 @@ export const formatVolume = (value, unit) => {
 export const formatHours = (value) => {
   return moment.duration(value).humanize();
 };
-
-const speedConverter = (value, unit) => {
-  switch (unit) {
-    case 'kmh':
-      return (value * 1.852).toFixed(2);
-    case 'mph':
-      return (value * 1.15078).toFixed(2);
-    case 'kn':
-    default:
-        return (value * 1).toFixed(2);
-  }
-};
-
-export const getConverter = (key) => {
-  switch (key) {
-    case 'speed':
-      return speedConverter;
-    default:
-      return function (value) {
-        return value;
-      }
-  }
-};
